# cmd/swag/docs/swagger.yaml
basePath: /v1/2024
definitions:
  model.LoginInput:
    properties:
      user_account:
        type: string
      user_password:
        type: string
    type: object
  model.RegisterInput:
    properties:
      verify_key:
        type: string
      verify_purpose:
        type: string
      verify_type:
        type: integer
    type: object
  model.UpdatePasswordRegisterInput:
    properties:
      password:
        type: string
      user_token:
        type: string
    type: object
  model.VerifyInput:
    properties:
      verify_code:
        type: string
      verify_key:
        type: string
    type: object
  response.ErrorResponseData:
    properties:
      code:
        description: mã trạng thái trả về
        type: integer
      detail:
        description: dữ liệu trả về
      message:
        description: thông điệp trả về
        type: string
    type: object
  response.ResponseData:
    properties:
      code:
        description: mã trạng thái trả về
        type: integer
      data:
        description: dữ liệu trả về
      message:
        description: thông điệp trả về
        type: string
    type: object
host: localhost:8002
info:
  contact:
    email: anhduynguyen1@gmail.com
    name: TEAM TIPSGO
    url: github.com/anhduynguyen1207/go-ecommerce-backend-go
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: github.com/anhduynguyen1207/go-ecommerce-backend-go
  title: API Documentation Ecommerce Backend SHOPDEVGO
  version: 1.0.0
paths:
  /user/login:
    post:
      consumes:
      - application/json
      description: User Login
      parameters:
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
      summary: User Login
      tags:
      - account management
  /user/register:
    post:
      consumes:
      - application/json
      description: When user is register send otp to email
      parameters:
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.RegisterInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
      summary: User Register
      tags:
      - account management
  /user/verify_account:
    post:
      consumes:
      - application/json
      description: Verify OTP Login By User
      parameters:
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.VerifyInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
      summary: Verify OTP Login By User
      tags:
      - account management
  /user/verify_update_pass_register:
    post:
      consumes:
      - application/json
      description: UpdatePasswordRegister
      parameters:
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.UpdatePasswordRegisterInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
      summary: UpdatePasswordRegister
      tags:
      - account management
swagger: "2.0"