# pkg/response/response.go
package response

import (
	"net/http"

	"github.com/gin-gonic/gin"
)

type ResponseData struct {
	Code    int         `json:"code"`    //mã trạng thái trả về
	Message string      `json:"message"` //thông điệp trả về
	Data    interface{} `json:"data"`    //dữ liệu trả về
}

type ErrorResponseData struct {
	Code    int         `json:"code"`    //mã trạng thái trả về
	Message string      `json:"message"` //thông điệp trả về
	Detail  interface{} `json:"detail"`  //dữ liệu trả về
}

// success response
func SuccessResponse(c *gin.Context, code int, data interface{}) {
	c.JSON(http.StatusOK, ResponseData{
		Code:    code,
		Message: msg[code],
		Data:    data,
	})
}

// error response
func ErrorResponse(c *gin.Context, code int, message string) {
	c.JSON(http.StatusOK, ResponseData{
		Code:    code,
		Message: msg[code],
		Data:    nil,
	})
}